requirements.txt
# Core Streamlit and web framework
streamlit>=1.28.0
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0

# Authentication
bcrypt>=4.0.0

# AI and ML
google-generativeai>=0.3.0

# Data visualization and analysis
plotly>=5.17.0
pandas>=2.1.0

# Audio processing
SpeechRecognition>=3.10.0
st-audiorec>=0.0.5

# HTTP requests
requests>=2.31.0

# Additional utilities
python-multipart>=0.0.6
Pillow>=10.0.0

# Production server (optional)
gunicorn>=21.2.0
